Establish shardingsphere-pluggable module (#4988)
* rename BaseShardingEngine to BasePrepareEngine

* rename BaseShardingEngine to BasePrepareEngine

* rename sharding-core-entry to shardingsphere-pluggable

* remove sql parser dialect import for shardingsphere-pluggable

* remove merge from pluggable module

* remove useless test cases

* remove specific route and rewrite from pluggable module

* remove sharding-core-common from pluggable module

* move SQLExecutionHook to shardingsphere-executor module

* remove sharding-core-execute from pluggable module

* move shardingsphere-pluggable to shardingsphere-underlying

* rename package
