+++ sharding-core/pom.xml
@@ -35,6 +35,5 @@
         <module>sharding-core-rewrite</module>
         <module>sharding-core-execute</module>
         <module>sharding-core-merge</module>
-        <module>sharding-core-entry</module>
     </modules>
 </project>
