+++ sharding-opentracing/src/test/java/org/apache/shardingsphere/opentracing/hook/OpenTracingSQLExecutionHookTest.java
@@ -22,8 +22,8 @@ import io.opentracing.ActiveSpan.Continuation;
 import io.opentracing.mock.MockSpan;
 import io.opentracing.tag.Tags;
 import org.apache.shardingsphere.opentracing.constant.ShardingTags;
-import org.apache.shardingsphere.sharding.execute.sql.hook.SPISQLExecutionHook;
-import org.apache.shardingsphere.sharding.execute.sql.hook.SQLExecutionHook;
+import org.apache.shardingsphere.underlying.executor.hook.SPISQLExecutionHook;
+import org.apache.shardingsphere.underlying.executor.hook.SQLExecutionHook;
 import org.apache.shardingsphere.spi.NewInstanceServiceLoader;
 import org.apache.shardingsphere.spi.database.metadata.DataSourceMetaData;
 import org.apache.shardingsphere.underlying.executor.engine.ExecutorDataMap;
