+++ sharding-orchestration/sharding-orchestration-center/sharding-orchestration-center-api/pom.xml
@@ -34,7 +34,7 @@
         </dependency>
         <dependency>
             <groupId>org.apache.shardingsphere</groupId>
-            <artifactId>sharding-core-entry</artifactId>
+            <artifactId>shardingsphere-pluggable</artifactId>
             <version>${project.version}</version>
         </dependency>
     </dependencies>
