+++ sharding-orchestration/sharding-orchestration-core/sharding-orchestration-core-common/pom.xml
@@ -29,6 +29,11 @@
     <dependencies>
         <dependency>
             <groupId>org.apache.shardingsphere</groupId>
+            <artifactId>sharding-core-common</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.shardingsphere</groupId>
             <artifactId>sharding-orchestration-core-configuration</artifactId>
             <version>${project.version}</version>
         </dependency>
