+++ shardingsphere-underlying/pom.xml
@@ -35,5 +35,6 @@
         <module>shardingsphere-rewrite</module>
         <module>shardingsphere-executor</module>
         <module>shardingsphere-merge</module>
+        <module>shardingsphere-pluggable</module>
     </modules>
 </project>
