+++ shardingsphere-underlying/shardingsphere-executor/src/main/java/org/apache/shardingsphere/underlying/executor/hook/SQLExecutionHook.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sharding.execute.sql.hook;
+package org.apache.shardingsphere.underlying.executor.hook;
 
 import org.apache.shardingsphere.spi.database.metadata.DataSourceMetaData;
